+++ previous-versions/common/src/sdk/models/shared/pulloperation.ts
@@ -3,7 +3,6 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import { DataType } from "./datatype";
 import { Expose } from "class-transformer";
 
 /**
@@ -84,11 +83,11 @@ export class PullOperation extends SpeakeasyBase {
     connectionId: string;
 
     /**
-     * Available Data types
+     * The data type you are requesting in a pull operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "dataType" })
-    dataType: DataType;
+    dataType: string;
 
     /**
      * A message about a transient or persistent error.