+++ previous-versions/common/RELEASES.md
@@ -691,3 +691,13 @@ Based on:
 - [typescript v0.39.0] previous-versions/common
 ### Releases
 - [NPM v0.39.0] https://www.npmjs.com/package/@codat/common/v/0.39.0 - previous-versions/common
+
+## 2023-12-04 01:30:55
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.125.0 (2.209.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.39.1] previous-versions/common
+### Releases
+- [NPM v0.39.1] https://www.npmjs.com/package/@codat/common/v/0.39.1 - previous-versions/common